replace notify-send with testNotifier
Since the Shellshock bash bugfix, notify-send is no longer a valid
identifier, and cannot be used in the test. Echo is valid, but not an
option (we want echo to be usable in the tests and in shunit). No choice
but to allow the "testNotifier" call as an option. A bit ugly, but
works.